// frontend/web/src/services/API/proserve-wb-packaging-api/models/ValidateComponentVersionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ValidateComponentVersionRequest
 */
export interface ValidateComponentVersionRequest {
    /**
     * The YAML definition of the new component version
     * @type {string}
     * @memberof ValidateComponentVersionRequest
     */
    componentVersionYamlDefinition: string;
}

/**
 * Check if a given object implements the ValidateComponentVersionRequest interface.
 */
export function instanceOfValidateComponentVersionRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentVersionYamlDefinition" in value;

    return isInstance;
}

export function ValidateComponentVersionRequestFromJSON(json: any): ValidateComponentVersionRequest {
    return ValidateComponentVersionRequestFromJSONTyped(json, false);
}

export function ValidateComponentVersionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ValidateComponentVersionRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentVersionYamlDefinition': json['componentVersionYamlDefinition'],
    };
}

export function ValidateComponentVersionRequestToJSON(value?: ValidateComponentVersionRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'componentVersionYamlDefinition': value.componentVersionYamlDefinition,
    };
}